Data Sources & Pools: view/function discovery + UX optimization (Tier 1)
Backend:
- Discover endpoint fetches pg_description table comments, discovers views and functions separately
- Lifecycle API returns view/function counts; Organization query includes views
- New GET /resources/functions endpoint for discovered SQL functions
- Replace SELECT * with explicit columns in datasource.ts; rename dead 'skipped' array to 'updated'
- Fix assignment_count SQL to return int instead of string

Frontend:
- Discovery phase: 5 summary cards (tables/views/columns/functions/last discovered), View Functions table with volatility badges, table_type indicator
- Organization phase: VIEW/TABLE badge, display_name/table_comment shown, inline edit with pencil icon
- Replace all 23 alert() calls with toast notifications (useToast)
- Fix 13 silent catch blocks with console.warn + toast.error
- Add empty state guidance for Discovery, Profiles, Credentials, Deployment phases
- Wrap all reactivate/remove operations with DangerConfirmModal (was missing for 5 operations)
- Fix PoolProfile.assignment_count type from string to number

Docs:
- Add data mining thin-slice engine plan + long-term vision reference docs
- Update CLAUDE.md and wishlist-features.md with doc pointers

+++ services/authz-api/src/routes/pool.ts
@@ -30,7 +30,7 @@ poolRouter.get('/profiles', async (_req, res) => {
   try {
     const result = await pool.query(`
       SELECT dp.*,
-        (SELECT count(*) FROM authz_db_pool_assignment da
+        (SELECT count(*)::int FROM authz_db_pool_assignment da
          WHERE da.profile_id = dp.profile_id AND da.is_active) AS assignment_count
       FROM authz_db_pool_profile dp
       ORDER BY dp.profile_id
